remove arm64 and fully support only open ISAs, also add support for riscv32

+++ volkrnl/includes/rtl/format.h
@@ -3,13 +3,13 @@
 
 #include <types.h>
 
-void rtl_print_dec(const char* prev, u64 dec, boolean shouldNewline);
-void rtl_print_hex(const char* prev, u64 dec, boolean shouldNewline);
+void rtl_print_dec(const char* prev, usize dec, boolean shouldNewline);
+void rtl_print_hex(const char* prev, usize dec, boolean shouldNewline);
 
 
 #ifdef _VA_ARG
 #include <stdarg.h>
-void rtl_format(char* buf, u64 len, const char* format, va_list list);
+void rtl_format(char* buf, usize len, const char* format, va_list list);
 #endif
