+++ volkrnl/includes/rtl/string.h
@@ -9,14 +9,14 @@
 #include <types.h>
 
 
-u64 rtl_get_cstring_len(const char* src);
-void rtl_copy_cstring(const char* src, char* dest, u64 len);
-u64 rtl_cstring_to_dec_ex(const char* src, u64 len, boolean shouldFail);
-u64 rtl_cstring_to_dec(const char* src, u64 len);
-void rtl_dec_to_cstring(const u64 number, char* src, u64 len);
-u64 rtl_cstring_to_hex_ex(const char* src, u64 len, boolean shouldFail);
-u64 rtl_cstring_to_hex(const char* src, u64 len);
-void rtl_hex_to_cstring(const u64 number, char* src, u64 len);
+usize rtl_get_cstring_len(const char* src);
+void rtl_copy_cstring(const char* src, char* dest, usize len);
+usize rtl_cstring_to_dec_ex(const char* src, usize len, boolean shouldFail);
+usize rtl_cstring_to_dec(const char* src, usize len);
+void rtl_dec_to_cstring(const usize number, char* src, usize len);
+usize rtl_cstring_to_hex_ex(const char* src, usize len, boolean shouldFail);
+usize rtl_cstring_to_hex(const char* src, usize len);
+void rtl_hex_to_cstring(const usize number, char* src, usize len);
